Добавил в исключения папку idea, вынес парамтерезированные тесты в отдельные классы, поправил структуры классов без параметризированных тестов
