+++ src/test/java/LionTest.java
@@ -2,39 +2,19 @@ import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 import com.example.*;
 import java.util.Arrays;
-import java.util.Collection;
-import org.junit.Before;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
 
-@RunWith(Parameterized.class)
+@RunWith(MockitoJUnitRunner.class)
 public class LionTest {
-
     private Lion lion;
-    private Feline felineMock;
 
-    private String sex;
-    private boolean expectedHasMane;
+ @Mock
+ Feline felineMock;
 
-    @Before
-    public void setUp() throws Exception {
-        felineMock = mock(Feline.class);
-    }
-
-    public LionTest(String sex, boolean expectedHasMane) {
-        this.sex = sex;
-        this.expectedHasMane = expectedHasMane;
-    }
-
-    // Параметры для теста
-    @Parameterized.Parameters
-    public static Collection<Object[]> data() {
-        return Arrays.asList(new Object[][] {
-                {"Самец", true},
-                {"Самка", false}
-        });
-    }
     @Test
     public void testInvalidSex() {
         try {
@@ -44,18 +24,13 @@ public class LionTest {
         }
     }
 
-    @Test
-    public void testDoesHaveMane() throws Exception {
-        lion = new Lion(sex, felineMock);
-        assertEquals(expectedHasMane, lion.doesHaveMane());
-    }
 
 
     @Test
     public void testGetKittens() {
         try {
             when(felineMock.getKittens()).thenReturn(1);
-            lion = new Lion(sex, felineMock);
+            lion = new Lion("Самец", felineMock);
             assertEquals(1, lion.getKittens());
         } catch (Exception e) {
             fail("Unexpected exception: " + e.getMessage());
@@ -65,7 +40,7 @@ public class LionTest {
     @Test
     public void testGetFood() throws Exception {
         when(felineMock.getFood("Хищник")).thenReturn(Arrays.asList("Животные", "Птицы", "Рыба"));
-        lion = new Lion(sex, felineMock);
+        lion = new Lion("Самка", felineMock);
         assertEquals(Arrays.asList("Животные", "Птицы", "Рыба"), lion.getFood());
     }
 }
