+++ src/test/java/ParametrizedLionTest.java
@@ -1,0 +1,43 @@
+import com.example.Feline;
+import com.example.Lion;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+
+import java.util.Arrays;
+import java.util.Collection;
+
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
+
+@RunWith(Parameterized.class)
+public class ParametrizedLionTest {
+    private Lion lion;
+    private Feline felineMock;
+    private String sex;
+    private boolean expectedHasMane;
+    @Before
+    public void setUp(){
+        felineMock = mock(Feline.class);
+    }
+
+    public ParametrizedLionTest(String sex, boolean expectedHasMane) {
+        this.sex = sex;
+        this.expectedHasMane = expectedHasMane;
+    }
+
+    @Parameterized.Parameters
+    public static Collection<Object[]> data() {
+        return Arrays.asList(new Object[][] {
+                {"Самец", true},
+                {"Самка", false}
+        });
+    }
+
+    @Test
+    public void testDoesHaveMane() throws Exception {
+        lion = new Lion(sex, felineMock);
+        assertEquals(expectedHasMane, lion.doesHaveMane());
+    }
+}
